extract BlocksEditor component out of BlocksInput
